Start Refactoring Enhancer Gene Connection plot section into separate UI component directory. Adds enum for cell encoding format and starts to move to typescript (#16)
* Start Refactoring Enhancer Gene Connection plot section into separate UI component directory. Adds enum for cell encoding format and starts to move to typescript

* Refactor EnhancerGeneSettings to TS and use render info pattern for form options
