+++ src/components/enhancer-genes/EnhancerGenesHelp.tsx
@@ -2,11 +2,7 @@ import React from 'react';
 import Typography from '@material-ui/core/Typography';
 
 const EnhancerRegionsHelp = React.memo(function EnhancerRegionsHelp() {
-  return (
-    <React.Fragment>
-      <Typography>...</Typography>
-    </React.Fragment>
-  );
+  return <Typography>...</Typography>;
 });
 
 export default EnhancerRegionsHelp;
